#15: Handle searches with few characters
Set a minimum search length of 3 characters. Instead of showing an error to the user, we ignore the search and tell them to add more characters to the search term. This helps make it feel less like they made a mistake and more like they just need to be more specific.

Also disabled the browser's autocomplete because it appears to mess up Edge and covers text.